# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "BillingResponse": {
        "additionalProperties": false,
        "properties": {
          "capabilities": {
            "$ref": "#/components/schemas/BillingCapabilitiesResponse"
          },
          "communication_templates": {
            "items": {
              "$ref": "#/components/schemas/CommunicationTemplateResponse"
            },
            "title": "Communication Templates",
            "type": "array"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/BillingItemResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "owner": {
            "$ref": "#/components/schemas/BillingOwnerResponse"
          },
          "pix_key": {
            "title": "Pix Key",
            "type": "string"
          },
          "pix_merchant_city": {
            "title": "Pix Merchant City",
            "type": "string"
          },
          "pix_merchant_name": {
            "title": "Pix Merchant Name",
            "type": "string"
          },
          "pix_needs_setup": {
            "title": "Pix Needs Setup",
            "type": "boolean"
          },
          "recipients": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ContactReferenceResponse"
                },
                {
                  "$ref": "#/components/schemas/ContactResponse"
                }
              ]
            },
            "title": "Recipients",
            "type": "array"
          },
          "reply_to": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ContactReferenceResponse"
                },
                {
                  "$ref": "#/components/schemas/ContactResponse"
                }
              ]
            },
            "title": "Reply To",
            "type": "array"
          },
          "stats": {
            "$ref": "#/components/schemas/BillingStatsResponse"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "uuid": {
            "title": "Uuid",
            "type": "string"
          }
        },
        "required": [
          "uuid",
          "name",
          "description",
          "pix_key",
          "pix_merchant_name",
          "pix_merchant_city",
          "owner",
          "items",
          "recipients",
          "reply_to",
          "communication_templates",
          "stats",
          "pix_needs_setup",
          "capabilities"
        ],
        "title": "BillingResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/billings/{billing_uuid}": {
      "delete": {
        "operationId": "delete_billing_api_v1_billings__billing_uuid__delete",
        "parameters": [
          {
            "in": "path",
            "name": "billing_uuid",
            "required": true,
            "schema": {
              "title": "Billing Uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Billing",
        "tags": [
          "billings"
        ]
      },
      "get": {
        "operationId": "get_billing_api_v1_billings__billing_uuid__get",
        "parameters": [
          {
            "in": "path",
            "name": "billing_uuid",
            "required": true,
            "schema": {
              "title": "Billing Uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Billing",
        "tags": [
          "billings"
        ]
      },
      "patch": {
        "operationId": "update_billing_api_v1_billings__billing_uuid__patch",
        "parameters": [
          {
            "in": "path",
            "name": "billing_uuid",
            "required": true,
            "schema": {
              "title": "Billing Uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BillingUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Billing",
        "tags": [
          "billings"
        ]
      }
    },
...
    "/api/v1/billings/{billing_uuid}/bills": {
      "get": {
        "operationId": "list_bills_api_v1_billings__billing_uuid__bills_get",
        "parameters": [
          {
            "in": "path",
            "name": "billing_uuid",
            "required": true,
            "schema": {
              "title": "Billing Uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Bills",
        "tags": [
          "bills"
        ]
      },
      "post": {
        "operationId": "create_bill_api_v1_billings__billing_uuid__bills_post",
        "parameters": [
          {
            "in": "path",
            "name": "billing_uuid",
            "required": true,
            "schema": {
              "title": "Billing Uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BillCreateRequest"
              }
            },
            "multipart/form-data": {
              "encoding": {
                "payload": {
                  "contentType": "application/json"
                }
              },
              "schema": {
                "$ref": "#/components/schemas/Body_create_bill_api_v1_billings__billing_uuid__bills_post",
                "properties": {
                  "payload": {
                    "$ref": "#/components/schemas/BillCreateRequest"
                  },
                  "receipt_files": {
                    "items": {
                      "format": "binary",
                      "type": "string"
                    },
                    "type": "array"
                  }
                },
                "required": [
                  "payload"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Create Bill",
        "tags": [
          "bills"
        ]
      }
    },
...
    "/api/v1/billings/{billing_uuid}/bills/{bill_uuid}/invoice": {
      "get": {
        "operationId": "download_invoice_api_v1_billings__billing_uuid__bills__bill_uuid__invoice_get",
        "parameters": [
          {
            "in": "path",
            "name": "billing_uuid",
            "required": true,
            "schema": {
              "title": "Billing Uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "bill_uuid",
            "required": true,
            "schema": {
              "title": "Bill Uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Invoice",
        "tags": [
          "bills"
        ]
      }
    },
...
    "/api/v1/billings/{billing_uuid}/bills/{bill_uuid}/recibo": {
      "get": {
        "operationId": "download_recibo_api_v1_billings__billing_uuid__bills__bill_uuid__recibo_get",
        "parameters": [
          {
            "in": "path",
            "name": "billing_uuid",
            "required": true,
            "schema": {
              "title": "Billing Uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "bill_uuid",
            "required": true,
            "schema": {
              "title": "Bill Uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Recibo",
        "tags": [
          "bills"
        ]
      }
    },
...
  }
}